Teams table + associations.
Added a table for team info.
This will act as the "parent" for the games and keep the readable team
info.
Therefore, the home_team and away_team team in table games are now
foreign keys referencing the teams table.

Everything compiles but to verify that this actually works I need to
actually put some data in the database.