Database call working!
Actual logical bug caused the database call not to work.
The problem was that the serialised data from the server was Vec<Team>
but the ui tried to deserialise it as Teams (HashMap<TeamId, Team>).
Now, it seems to instantly work.
I removed the mock_data entirely, but without group games, nothing is
currently rendered.