Correct game ids from database
db/src/models.rs is one of the few places where values are represented
with primitive values (i32) and the newtype pattern is not available.
There, Pred id got mixed up with the game id.
I think this is a validation of the consistent use of newtypes in the
code base since it makes this class of bugs unrepresentable.

Fixes: #22