WIP: Max cap for GoalCount
I have this insane notion that people will break the UI the only way
they can: i.e. by overflowing the goal count when entering the score,
like:
EGY 1 - u32::MAX BRA

Therefore this commit removes From/Into u32 impls for GoalCount and
instead adds a fallible constructor which checks against a
GoalCount::MAX value (999 currently). This needed changes in a _lot_ of
places but is now fixed. There seems to be a ui issue though.
Committing now so that I can revert to before the change and see if this
change caused it.