Fixed inf loop fetch
The issue was that the display name was included in the AuthState enum
variant for convenience (got it with the same API call).
The authstate was accessed in the ProtectedRoute, which then re-rendered
on setting the display name, which recursed.

Decoupling display name from auth seems to have fixed it.